# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal Berita",
        "description": "Portal Berita API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "Staging server"
        },
        {
            "url": "http://example.com",
            "description": "Production server"
        }
    ],
    "paths": {
        "/api/anggotas": {
            "get": {
                "tags": [
                    "Anggota"
                ],
                "summary": "Menampilkan semua data anggota",
                "operationId": "5ffb4d5912a8241dc26c4e4fd641fe22",
                "responses": {
                    "200": {
                        "description": "Daftar anggota berhasil diambil"
                    },
                    "500": {
                        "description": "Terjadi kesalahan saat mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Anggota"
                ],
                "summary": "Menyimpan data anggota baru",
                "operationId": "6dd3d15df102497309a4d44ac76c64f4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "nomor_telepon",
                                    "email",
                                    "tanggal_bergabung"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Budi"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "budi@example.com"
                                    },
                                    "tanggal_bergabung": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Anggota berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Terjadi kesalahan"
                    }
                }
            }
        },
        "/api/anggotas/{id}": {
            "get": {
                "tags": [
                    "Anggota"
                ],
                "summary": "Menampilkan detail anggota",
                "operationId": "30c83ba97ca3eaaaf415f7941dedcbd4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID anggota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail anggota"
                    },
                    "404": {
                        "description": "Anggota tidak ditemukan"
                    },
                    "500": {
                        "description": "Terjadi kesalahan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Anggota"
                ],
                "summary": "Memperbarui data anggota",
                "operationId": "0095def83fa0d5e0b427f0ea8c8d10d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID anggota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Andi"
                                    },
                                    "nomor_telepon": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "andi@example.com"
                                    },
                                    "tanggal_bergabung": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Anggota berhasil diperbarui"
                    },
                    "404": {
                        "description": "Anggota tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Terjadi kesalahan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Anggota"
                ],
                "summary": "Menghapus data anggota",
                "operationId": "ead3baf3bbea4c11b65573f88c5dfad2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID anggota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anggota berhasil dihapus"
                    },
                    "404": {
                        "description": "Anggota tidak ditemukan"
                    },
                    "500": {
                        "description": "Terjadi kesalahan"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}